requirements.txt
pylint
GitPython